refactor(app_runner): rename the resource-dataset spec module, drop an unused alias
bfabric_resource_dataset.py was the only input spec module missing the
_spec suffix its eight siblings carry; rename it and update the src,
test, docs (autopydantic_model directive) and basedpyright baseline
references. The 'bfabric_resource_dataset' YAML discriminator value is
unchanged, so specs on disk keep working.

Also drop BfabricAnnotationSpecField, which had no reference anywhere in
the repo.
